requirements.txt
pandas
numpy
scikit-learn
joblib
pytest
dvc[gs]
google-cloud-storage
fastapi
uvicorn
pydantic
opentelemetry-exporter-gcp-trace